Seperate Settings's submenus
