+++ CodeHub/Views/Settings/GeneralSettingsView.xaml.cs
@@ -1,4 +1,4 @@
-﻿using CodeHub.ViewModels;
+﻿using CodeHub.ViewModels.Settings;
 using Windows.UI.Xaml;
 
 
@@ -6,12 +6,14 @@ namespace CodeHub.Views
 {
     public sealed partial class GeneralSettingsView : SettingsDetailPageBase
     {
-        private SettingsViewModel ViewModel;
+        private GeneralSettingsViewModel ViewModel;
+
         public GeneralSettingsView()
         {
             this.InitializeComponent();
 
-            ViewModel = new SettingsViewModel();
+            ViewModel = new GeneralSettingsViewModel();
+
             this.DataContext = ViewModel;
         }
         private void OnCurrentStateChanged(object sender, VisualStateChangedEventArgs e)
